--- defi-alerts/src/lib/auth.ts
@@ -1,5 +1,0 @@
-// File: src/lib/auth.ts
-export function isLoggedIn(): boolean {
-  if (typeof window === 'undefined') return false;
-  return localStorage.getItem('isLoggedIn') === 'true';
-}
